+++ .github/patch-types.php
@@ -46,6 +46,7 @@ foreach ($loader->getClassMap() as $class => $file) {
         case false !== strpos($file, '/src/Symfony/Component/DependencyInjection/Tests/Fixtures/php/'):
         case false !== strpos($file, '/src/Symfony/Component/DependencyInjection/Tests/Fixtures/TestServiceSubscriberIntersectionWithTrait.php'):
         case false !== strpos($file, '/src/Symfony/Component/ErrorHandler/Tests/Fixtures/'):
+        case false !== strpos($file, '/src/Symfony/Component/HttpClient/Internal/'):
         case false !== strpos($file, '/src/Symfony/Component/Form/Tests/Fixtures/Answer.php'):
         case false !== strpos($file, '/src/Symfony/Component/Form/Tests/Fixtures/Number.php'):
         case false !== strpos($file, '/src/Symfony/Component/Form/Tests/Fixtures/Suit.php'):