+++ src/Symfony/Component/HttpClient/CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
  * Allow mocking `start_time` info in `MockResponse`
  * Add `MockResponse::fromFile()` and `JsonMockResponse::fromFile()` methods to help using fixtures files
  * Add `ThrottlingHttpClient` to enable limiting the number of requests within a certain period
+ * Add support for amphp/http-client v5 on PHP 8.4+
 
 7.0
 ---